save site non fonctionel
raccorde le code et les page att

+++ app/routes/cars.py
@@ -3,7 +3,8 @@ from typing import Annotated
 from fastapi import APIRouter, HTTPException, status, Form, Depends
 from fastapi.responses import RedirectResponse
 from pydantic import ValidationError
-from app.schemas import cars, UserSchema # Assuming you have a Car and UserSchema schema
+from app.schemas.cars import Car # Assuming you have a Car and UserSchema schema
+from app.schemas.UserSchema import UserSchema
 import app.services.cars as service  # Assuming you have a service module for cars
 from app.login_manager import login_manager
 from fastapi.templating import Jinja2Templates
@@ -15,18 +16,9 @@ templates = Jinja2Templates(directory="templates")
 
 @router.get('/')
 def get_all_cars(request: Request, user: UserSchema = Depends(login_manager.optional)):
-    if user is not None:
-        if user.role == "admin":
-            cars = service.get_all_cars()
-        else:
-            cars = service.get_public_cars()
-            cars_self = service.get_own_cars(user)
-            cars = cars + cars_self
-    else:
-        cars = service.get_public_cars()
     return templates.TemplateResponse(
-        "all_cars.html",
-        context={'request': request, 'current_user': user, 'cars': cars}
+            "front_page.html",
+            context={'request': request}
     )
 
 @router.get('/new')
@@ -41,10 +33,10 @@ def get_car(request: Request, user: UserSchema = Depends(login_manager)):
             status_code=status.HTTP_401_UNAUTHORIZED,
             detail="You are blocked."
         )
-    elif user.role != "admin":
+    elif user.role != "vendeur":
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="Only admins can use this page"
+            detail="seul les vendeurs peuvent utiliser cette page"
         )
     return templates.TemplateResponse(
         "new_car.html",
@@ -52,52 +44,64 @@ def get_car(request: Request, user: UserSchema = Depends(login_manager)):
     )
 
 @router.post('/new')
-def create_new_car(make: Annotated[str, Form()], model: Annotated[str, Form()], year: Annotated[int, Form()], owner: Annotated[str, Form()] = None):
-    if owner is not None:
-        if get_user_by_email(owner) is None:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="No owner with this email"
-            )
+def create_new_car(make: Annotated[str, Form()], model: Annotated[str, Form()],color:Annotated[str,Form()],max_speed: Annotated[int,Form()], mileage: Annotated[int, Form()],average_consumption: Annotated[int,Form()],user: UserSchema = Depends(login_manager),price_sell:Annotated[float, Form()]=0,price_rent: Annotated[float,Form()]=0):
+    if price_sell==0 and price_rent==0:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="please input a price for rent or sell",
+        )
     new_car_data = {
         "id": str(uuid4()),
-        "make": make,
+        "brand": make,
         "model": model,
-        "year": year,
-        "owner": owner,
+        "color": color,
+        "max_speed": max_speed,
+        "mileage": mileage,
+        "average_consumption": average_consumption,
+        "price_sell": price_sell,
+        "price_rent": price_rent,
+        "owner_email": user.email,
+        "sell":True,
+        "rent":True
+        
     }
     try:
-        new_car_test = cars(**new_car_data)
+        new_car_test = Car(**new_car_data)
     except ValidationError:
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Invalid make, model, or year for the car.",
+            detail="Invalid make, model, or other for the car.",
         )
     service.save_car(new_car_test)
     return RedirectResponse(url="/cars/", status_code=302)
 
 @router.post('/modify')
-def modify_car(id: Annotated[str, Form()], make: Annotated[str, Form()], model: Annotated[str, Form()], year: Annotated[int, Form()], owner: Annotated[str, Form()] = None):
+def modify_car(id: Annotated[str, Form()],make: Annotated[str, Form()], model: Annotated[str, Form()],color:Annotated[str,Form()],max_speed: Annotated[int,Form()], mileage: Annotated[int, Form()],average_consumption: Annotated[int,Form()],user: UserSchema = Depends(login_manager),price_sell:Annotated[float, Form()]=0,price_rent: Annotated[float,Form()]=0):
     if not service.is_car_exist(id):
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Car ID not found"
         )
-    if owner is not None:
-        if get_user_by_email(owner) is None:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="No owner with this email"
-            )
+    if user.seller!=True:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="only seller can modify cars")
     new_car_data = {
         "id": id,
-        "make": make,
-        "model": model,
-        "year": year,
-        "owner": owner,
+        "brand": make, #supr?
+        "model": model, #supr?
+        "color": color, #supr?
+        "max_speed": max_speed, #supr?
+        "mileage": mileage,
+        "average_consumption": average_consumption, #supr?
+        "price_sell": price_sell,
+        "price_rent": price_rent,
+        "owner_email": user.email, 
+        "sell":(price_sell!=0),
+        "rent":(price_rent!=0),
     }
     try:
-        new_car_test = cars(**new_car_data)
+        new_car_test = Car(**new_car_data)
     except ValidationError:
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
@@ -107,16 +111,36 @@ def modify_car(id: Annotated[str, Form()], make: Annotated[str, Form()], model: 
     return RedirectResponse(url="/cars/", status_code=302)
 
 @router.post('/delete')
-def delete_car(id: Annotated[str, Form()]):
+def deletecar(id: Annotated[str, Form()], user: UserSchema = Depends(login_manager.optional)):
+    if user is None:
+            return RedirectResponse(url="/users/login", status_code=302)
+
+    elif user.blocked==True:
+        return HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="blocked."
+        )
+    elif user.role!="admin" or user.seller!=True:
+        return HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="only admins and sellers can use this page"
+        )
+   
     if not service.is_car_exist(id):
-        raise HTTPException(
+        return HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
-            detail="Car ID not found"
+                detail=" id not found"
+        )
+    car_to_delete=service.get_car_by_id(id)
+    if car_to_delete.owner_email!=user.email or car_to_delete.rent_owner_email:
+        return HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="you cannot delete a car that is rented or not yours"
         )
     service.delete_car_by_id(id)
     return RedirectResponse(url="/cars/", status_code=302)
 
-@router.post('/sell')
+"""@router.post('/sell')
 def sell_car(id: Annotated[str, Form()], user: UserSchema = Depends(login_manager)):
     car = service.get_car_by_id(id)
     if car is None:
@@ -124,20 +148,53 @@ def sell_car(id: Annotated[str, Form()], user: UserSchema = Depends(login_manage
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Car not found"
         )
-    if user.email != car.owner and user.role != "admin":
+    if user.email != car.owner_email and user.role != "admin":
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
             detail="You are not the owner of this car."
         )
-    if car.status == "en vente":
+    if car.sell == True:
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
             detail="This car is already listed for sale."
         )
-    modified_car = car.copy(update={"status": "en vente"})
-    service.modify_car_by_id(car.id, modified_car)
-    return RedirectResponse(url="/users/profile", status_code=302)
-
+    elif car.rent_owner_email is not None:
+         raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="This car is already rented by someone else."
+        )
+    service.sell_car(car.id, modified_car)
+    return RedirectResponse(url="/users/profile", status_code=302)"""
+"""@router.post('/sell')
+def sell_book(id: Annotated[str, Form()],user: UserSchema = Depends(login_manager)):
+    book=service.get_book_by_id(id)
+    if book is None:
+        return HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+                detail=" book not found"
+        )
+    if user.email!=book.owner_email and user.role!="admin":
+        return HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="that book is not yours."
+        )
+    elif book.status=="en vente":
+        return HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="already selling this book"
+        )
+    else:
+        modified_book={
+            "Author":book.Author,
+            "Editor":book.Editor,
+            "id":book.id,
+            "name":book.name,
+            "Prix":book.Prix,
+            "Owner":book.owner_email,
+            "status":"en vente",
+        }
+        service.modify_book_by_id(book.id,modified_book)
+        return RedirectResponse(url="/users/profile", status_code=302)
 @router.post('/unsell')
 def retire_car_from_sale(id: Annotated[str, Form()], user: UserSchema = Depends(login_manager)):
     car = service.get_car_by_id(id)
@@ -158,7 +215,7 @@ def retire_car_from_sale(id: Annotated[str, Form()], user: UserSchema = Depends(
         )
     modified_car = car.copy(update={"status": "privé"})
     service.modify_car_by_id(car.id, modified_car)
-    return RedirectResponse(url="/users/profile", status_code=302)
+    return RedirectResponse(url="/users/profile", status_code=302)"""
 
 @router.post('/buy')
 def buy_car(id: Annotated[str, Form()], user: UserSchema = Depends(login_manager)):
@@ -168,16 +225,81 @@ def buy_car(id: Annotated[str, Form()], user: UserSchema = Depends(login_manager
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Car not found"
         )
-    if car.status != "en vente":
+    if car.sell !=True:
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
             detail="This car is not listed for sale."
         )
-    modified_car = car.copy(update={"owner": user.email, "status": "privé"})
-    service.modify_car_by_id(car.id, modified_car)
+    elif car.rent_owner_email is not None:
+        return HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="you buy a car that is rented"
+        )
+    new_car_data = {
+        "id": id,
+        "brand": car.make, #supr?
+        "model": car.model, #supr?
+        "color": car.color, #supr?
+        "max_speed": car.max_speed, #supr?
+        "mileage": car.mileage,
+        "average_consumption": car.average_consumption, #supr?
+        "price_sell": 0,
+        "price_rent": 0,
+        "owner_email": user.email, 
+        "sell":False,
+        "rent":False,
+    }
+    service.modify_car_by_id(car.id,new_car_data)
     return RedirectResponse(url="/cars/", status_code=302)
 
-@router.post("/change_owner")
+@router.post('/rent')
+def buy_car(id: Annotated[str, Form()], user: UserSchema = Depends(login_manager)):
+    car = service.get_car_by_id(id)
+    if car is None:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Car not found"
+        )
+    if car.rent !=True:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="This car is not listed for rent."
+        )
+    elif car.rent_owner_email is not None:
+        return HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="you rent a car that is already rented"
+        )
+    if user.seller==True:
+        return HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="only admins and buyers can use this page"
+        )
+    new_car_data = {
+        "id": id,
+        "brand": car.make, #supr?
+        "model": car.model, #supr?
+        "color": car.color, #supr?
+        "max_speed": car.max_speed, #supr?
+        "mileage": car.mileage,
+        "average_consumption": car.average_consumption, #supr?
+        "price_sell": car.sell,
+        "price_rent": car.rent,
+        "owner_email": user.email, 
+        "sell":car.sell,
+        "rent":False,
+        "rent_owner_email":user.email
+    }
+    try:
+        new_car_test = Car(**new_car_data)
+    except ValidationError:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Invalid make, model, or year for the car.",
+        )
+    service.modify_car_by_id(car.id,new_car_data,True)
+    return RedirectResponse(url="/cars/", status_code=302)
+"""@router.post("/change_owner")
 def change_car_owner(id: Annotated[str, Form()], new_owner: Annotated[str, Form()], user: UserSchema = Depends(login_manager)):
     car = service.get_car_by_id(id)
     if car is None:
@@ -197,9 +319,9 @@ def change_car_owner(id: Annotated[str, Form()], new_owner: Annotated[str, Form(
         )
     modified_car = car.copy(update={"owner": new_owner})
     service.modify_car_by_id(car.id, modified_car)
-    return RedirectResponse(url="/cars/", status_code=302)
+    return RedirectResponse(url="/cars/", status_code=302)"""
 
-@router.post('/rent')
+@router.post('/unrent')
 def rent_car(id: Annotated[str, Form()], user: UserSchema = Depends(login_manager)):
     car = service.get_car_by_id(id)
     if car is None:
@@ -207,17 +329,42 @@ def rent_car(id: Annotated[str, Form()], user: UserSchema = Depends(login_manage
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Car not found"
         )
-    if car.status != "privé":
+    if car.rent_owner_email is None:
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
-            detail="This car is not available for rent."
+            detail="This car is not rented."
         )
-    modified_car = car.copy(update={"status": "en location"})
-    service.modify_car_by_id(car.id, modified_car)
+    if car.rent_owner_email != user.email:
+     raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="This car is not rented by you."
+        ) 
+    new_car_data = {
+        "id": id,
+        "brand": car.make, #supr?
+        "model": car.model, #supr?
+        "color": car.color, #supr?
+        "max_speed": car.max_speed, #supr?
+        "mileage": car.mileage,
+        "average_consumption": car.average_consumption, #supr?
+        "price_sell": car.sell,
+        "price_rent": car.rent,
+        "owner_email": user.email, 
+        "sell":car.sell,
+        "rent":True,
+        "rent_owner_email": None
+    }
+    try:
+        new_car_test = Car(**new_car_data)
+    except ValidationError:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Invalid make, model, or year for the car.",
+        )
+    service.modify_car_by_id(car.id, new_car_data,True)
     return RedirectResponse(url="/cars/", status_code=302)
 
 
 @router.post('/search')
 def search(search_therm: Annotated[str,Form()]):
     service.search(search_therm)
-
